Select a single property.
Created initial project with `.gitignore`, `.npmignore`, `package.json`.
All set to build on Travis CI.

Created initial test to see if we can select a single object.

Using a `utility/test` to run the tests so that on Travis CI, I can
minify the source and run the tests a second time against the minified
source.

Brought the `utility/sizes` utility over from **Timezone**. It is run as
part of the Travis CI build. Inquiry will be a MicroJS library.

    File                             Raw         Min         Zip     Min/Zip
    index.js                  1002/00.98   440/00.43   476/00.46   310/00.30

Closes #1.