Implemented sub-queries.
Implemented sub-queries using square brackets to indicate a sub-query
predicate. Curly braces are now used to indicate a JavaScript predicate.

The above is a better description and distinction than the ones given in
the `README.md`. At some point before the next release the headings for
the `README.md` ought to be changed to **JavaScript Predicates** and
**Sub-Query Predicates**. StackOverflow has a nice discussion of the
meaning of the word [predicate in computer
science](http://stackoverflow.com/questions/3230944/what-does-predicate-mean-in-the-context-of-computer-science).

Closes #15.