Remove unnecessary variable.
Remove a variable that simply mirrors a function parameter for a gain of
5 bytes.

An easy way to get save space is to look for variables that can be
reused. I like to use the `$` variable to accpet the return value of a
regular expression because it makes me think of Perl. Thoughts Perl
makes me happy, so I like to think of Perl. It also means that I get to
reuse the `$` variable. To me it always means the values of the last
regular expression that was run.

When I want to minzip further, I look for other candidates for variable
reuse. Did I declare `i`, `j`, and `k`, when the loops are subsequent
and not nested? If so, then just reuse `i`.

This is going to anger some nerds who believe that a variable is sacred
and exists to serve only one purpose. Obviously, the objectives here are
not in line with the objectives of that princple. This is a trade off
and I'm trading clarity for bytes.

However, you can make be verbose in your local variable naming, because
they get rewritten by the minifier anyway. If you wanted to use a
variable for a "counter", then for "loop limit", name it
`counterThenLoopLimit`. It might help you remember that the variable is
reused.

See #46.

File                             Raw         Min         Zip     Min/Zip
index.js                  5674/05.54  2121/02.07  1872/01.83  1016/00.99