Replace variable with recalculation.
Gained 2 bytes by replacing a variable with a recalculation of the
value. Even though the calculation is longer and more verbose than the
variable, the removal of the variable shortens the code because gzip is
going to combine the repeated statement, but the variable can't be
reduced. Each variable you add increased the minzipped size.

See #46.

File                             Raw         Min         Zip     Min/Zip
index.js                  5616/05.48  2122/02.07  1854/01.81  1011/00.99