Create array instead of setting length to zero.
We need to have the creation statement to create the array. We may as
wel reuse it each time in the loop. That way, we don't create the array
at the start of the function, only when needed.

It removes the assignment to length. The assignment to length will not
be well minified, because property names cannot be rewritten. When we
want to minify, we look to avoid object properties.

Save 2 more bytes.

See #46.

File                             Raw         Min         Zip     Min/Zip
index.js                  5448/05.32  2109/02.06  1833/01.79  1007/00.98