Shorten property names.
Object property names cannot be minified. The minifier will shorten
local variable names, renaming them `a`, `b`, `c`, etc. It can't change
property names. It cannot assume anything about the scope of object
property names.

We can shorten them ourselves, though. It obfuscates a bit, but in our
case, we're assigning the properties to local variables in order to work
with them, so it is not too difficult to follow along.

With some projects, I've found that upon reducing a large function
signature into an `options` object, the minzipped size of the file grew
dramatically. Although the signature was large, it was consistent and
repeated through the source. It was rewritten so that all of the
parameters were single characters, and the same single characters so
that gzip could combine them.

Here, we've renamed `object` to `o` and `path` to `p`. We've saved 12
bytes. We should have enough room to implement whitespace insensitivity,
but there is some more minification we can do.

See #46.

File                             Raw         Min         Zip     Min/Zip
index.js                  5363/05.24  2024/01.98  1826/01.78   994/00.97