Whitespace for clarity.
Added 2 bytes. Got whitespace for the sake of clarity.

Switching to double-quoted strings to delineate path parts with special
characters was necessary because with backtick escaping, the slashes
were the delimiters. Having whitespace for clarity means we need both
quotes and escapes. That's fine. We'll just reused double-quotes
strings. This reduces size, because the regular expression is already in
the code, so we've already paid for it to gzip.

We can use JSON.parse to convert the string we extract, which won't
minify at all, but it is still less verbose than doing the escaping
ourselves. Plus, we remove the fixup that rewrote the `%` sign because
of JSON pointer.

We spend our savings on our calls to `trim()`. We use `trim()`
consistently instead of adding to the regular expressions. It will not
get minified, but it will be a consistent repetition for gzip.

While writing this, I found that I'm tracking the path in the
interpreter incorrectly, and that I'm never using it. I'm going to leave
it in there for now, until I figure out which way I want to go with it.

Closes #46.

File                             Raw         Min         Zip     Min/Zip
index.js                  5531/05.40  2059/02.01  1788/01.75   980/00.96